pyproject.toml
[project]
name = "django-ninja-starter"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-ninja>=1.4.3",
]
